% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.samples.R
\name{its.apply}
\alias{its.apply}
\title{massits time series functions}
\usage{
its.apply(m, fun, bands_params = NULL, bands = its.bands())
}
\arguments{
\item{m}{A valid massits tibble}

\item{fun}{A valid function with one argument to be applied on each attribute
(its first argument).}

\item{bands_params}{An optional parameter to be passed as \code{fun} second argument.
If informed, it must have \code{1} or the same number of bands.
(Default \code{NULL})}

\item{bands}{A string vector indicating those fields (bands) to which
function \code{fun} will be applied. All non selected fields
will remain unchanged (Default \code{its.bands()}).}
}
\value{
Massits table
}
\description{
Applies a function to some massits tibble's fields
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
