% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.samples.R
\name{its.apply_na}
\alias{its.apply_na}
\title{massits time series functions}
\usage{
its.apply_na(m, na_values = -3000, bands = its.bands())
}
\arguments{
\item{m}{A valid massits tibble}

\item{na_values}{The missing value to be substituted by \code{NA}
(Default \code{-3000})}

\item{bands}{A string vector indicating those fields (bands)
to be interpolated. All non selected fields
will remain unchanged (Default \code{its.bands()}).}
}
\value{
Massits table
}
\description{
Substitute some missing values by \code{NA} in massits tibble's fields.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
