% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.utils.R
\name{its.bands}
\alias{its.bands}
\title{massits utils functions}
\usage{
its.bands(m)
}
\arguments{
\item{m}{A valid massits tibble}
}
\value{
A string vector with all bands
}
\description{
Return the bands of massits \code{m} data.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
