% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.R
\docType{data}
\name{its.class.cols}
\alias{its.class.cols}
\alias{its.pred.cols}
\title{massits utils globals}
\format{An object of class \code{character} of length 6.}
\usage{
its.pred.cols
}
\description{
Massits features tibble column names
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{datasets}
