% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.features.R
\name{its.feat.apply}
\alias{its.feat.apply}
\title{massits features functions}
\usage{
its.feat.apply(f, fun)
}
\arguments{
\item{f}{A valid massits measures tibble}

\item{fun}{A valid function with one argument to be applied on each measure attribute.}
}
\value{
Massits feature table
}
\description{
Applies a function to all feature tibble's attributes columns.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
