% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.samples.R
\docType{data}
\name{its.feat.cols}
\alias{its.feat.cols}
\title{massits utils globals}
\format{An object of class \code{character} of length 1.}
\usage{
its.feat.cols
}
\description{
Massits features tibble column names
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{datasets}
