% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.features.R
\name{its.feat.create_folds}
\alias{its.feat.create_folds}
\title{massits features functions}
\usage{
its.feat.create_folds(f, cross)
}
\arguments{
\item{f}{A valid massits features tibble}

\item{cross}{Number of partitions to be used in cross validation}
}
\value{
Function to select a given partition
}
\description{
Return a function to select a partition of a massits features tibble.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
