% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.features.R
\name{its.feat.drop_na}
\alias{its.feat.drop_na}
\title{massits utils functions}
\usage{
its.feat.drop_na(f)
}
\arguments{
\item{f}{A valid massits features tibble}
}
\value{
Massits features tibble
}
\description{
Drop all rows with NA measures in \code{f} data.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
