% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.utils.R
\name{its.feat.length}
\alias{its.feat.length}
\title{massits utils functions}
\usage{
its.feat.length(f)
}
\arguments{
\item{f}{A valid massits features tibble}
}
\value{
Numeric
}
\description{
Return the length of measures in \code{f} data.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
