% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.features.R
\name{its.feat}
\alias{its.feat}
\title{massits features functions}
\usage{
its.feat(m, bands, time_break = its.t_break("2000-09-01", "12 months"),
  measure_id = NULL, drop_na = TRUE, cores = 1)
}
\arguments{
\item{m}{A valid massits tibble}

\item{bands}{A valid string vector of band names}

\item{time_break}{A numeric vector indicating the segments of time series to be breaked.
This vector can be computed in \code{\link{its.t_break}} function
(Default \code{its.t_break("2000-09-01", "12 months")}).}

\item{measure_id}{A numeric vector indicating the measure sequences to compose the feature tibble}

\item{drop_na}{Logical indicating wether to drop rows with any \code{NA} measured values (Default \code{TRUE}).}

\item{cores}{A \code{multidplyr} argument to enable multithread processing.}
}
\value{
Massits feature tibble
}
\description{
Transforms a massits tibble as an massits feature tibble
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
