% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.formula.R
\name{its.formula.linear}
\alias{its.formula.linear}
\title{massits formula functions}
\usage{
its.formula.linear(f = NULL)
}
\arguments{
\item{f}{A massits features tibble to compose the formula expression.}
}
\value{
Formula
}
\description{
Create a formula to be used in some Machine Learning models
              with predictors from massits features tibble \code{f} atributes.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
