% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.interp.R
\name{its.interp.k}
\alias{its.interp.k}
\title{massits interpolation functions}
\usage{
its.interp.k(m, interp = c("spline", "last", "linear"), bands = its.bands(),
  k = 23)
}
\arguments{
\item{m}{A valid massits tibble}

\item{interp}{A string indicating what interpolation to use.
Can assume one of the following values:
\code{"last"}, \code{"linear"}, and \code{"spline"}
(Default \code{"spline"}).}

\item{bands}{A string vector indicating those fields (bands)
to be interpolated. All non selected fields
will remain unchanged (Default \code{its.bands()}).}

\item{k}{How many points to resample each time series.}
}
\value{
Massits table
}
\description{
Performs a resampling onsome massits tibble's columns.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
