% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.ml.R
\name{its.ml.create_predict}
\alias{its.ml.create_predict}
\title{massits features functions}
\usage{
its.ml.create_predict(f, ml_model = its.ml.model.svm_radial(),
  summation = c("probabilities", "entropy", "rentropy", "none"))
}
\arguments{
\item{f}{A valid massits measures tibble}

\item{ml_model}{A valid machine learning model function}

\item{summation}{String value indicating a summation regarding the prediction}
}
\value{
Massits predicted tibble
}
\description{
This function returns a predict function to be applied on data to be classified.
              Besides the prediction, this function can return a information regarding the
              classification quality. The returned information, defined in \code{summation}
              parameter can be \code{"probabilities"}, the reliability of each reference
              label; \code{"entropy"}, the Shanon entropy given by
              \eqn{e=-\sum(p_i\log{p_i})}{e = -sum(p * log p)}; \code{"rentropy"}, the relative
              Shanon entropy, given by \eqn{re = \frac{e}{\log{k}}}{re = e / (log k)};
              and \code{"none"}, nothing is returned.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
