% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.ml.R
\name{its.ml.model.svm_radial}
\alias{its.ml.model.svm_radial}
\title{massits machine learning functions}
\usage{
its.ml.model.svm_radial(f = NULL, formula = its.formula.linear(),
  gamma = 1/its.feat.length(f), cost = 1)
}
\arguments{
\item{f}{A valid massits features tibble}

\item{formula}{A valid massits formula to be used in classification model}

\item{gamma}{Radial kernel parameter (\eqn{e^{(-gamma*|u-v|^2)}}{exp(-gamma*|u-v|^2)})}

\item{cost}{Cost of constraint violation}
}
\value{
SVM Model
}
\description{
SVM model. This function is a wraper to the \code{\link{svm}}
              function of \code{e1071} package.
              If no massits features tibble is informed, an enclosed function is returned.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
