% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.raster.R
\name{its.raster.chunk}
\alias{its.raster.chunk}
\alias{its.raster.next_chunk}
\title{massits raster functions}
\usage{
its.raster.next_chunk(chunk.tb)
}
\arguments{
\item{chunk.tb}{A valid massits tibble created by \code{\link{its.raster}} function.
This tibble must have an \code{its_raster} object in its attribute.}
}
\value{
Massits tibble with all data of current chunk.
}
\description{
Loads a chunk of data from a valid massits raster. If there is no
              more chunk to load from rasters, throws an exception.
              Check if there are chunks to be read with \code{\link{its.raster.end}} function.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
