% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.raster.R
\name{its.raster.end}
\alias{its.raster.end}
\title{massits raster functions}
\usage{
its.raster.end(chunk.tb)
}
\arguments{
\item{chunk.tb}{A valid massits tibble created by \code{\link{its.raster}} function.
This tibble must have an \code{its_raster} object in its attribute.}
}
\value{
Logical
}
\description{
Check if there are chunks to be loaded in \code{r} massits raster.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
