% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.raster.R
\name{its.raster}
\alias{its.raster}
\title{massits raster functions}
\usage{
its.raster(bands = list(evi = "~/Downloads/sinop-crop-evi.tif", ndvi =
  "~/Downloads/sinop-crop-ndvi.tif"), chunk_size = 4900)
}
\arguments{
\item{bands}{A named list with all raster files, e.g. \code{band_name="file_path"}.}

\item{chunk_size}{A raw number or a function informing the number of pixels to
be computed. The function receives as its argument the raster brick object.
This number is rounded to rows (Default \code{4900} \eqn{(70x70)}).}
}
\value{
Massits tibble with an \code{its_raster} object in its attributes.
}
\description{
Creates a new massits raster data from a list of files.
              The object can be used in its.raster function family.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
