% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.raster.R
\name{its.raster.save_chunk}
\alias{its.raster.save_chunk}
\title{massits raster functions}
\usage{
its.raster.save_chunk(chunk.tb, file = "~/Downloads/test.tif",
  overwrite = FALSE, save_bylayer = TRUE, save_bands = its.bands())
}
\arguments{
\item{chunk.tb}{A valid massits predicted tibble.}

\item{file}{A file path to be used as base name of files to be saved.}

\item{overwrite}{Logical indicating if file can be overitten (Default \code{FALSE}).}

\item{save_bylayer}{Save each time period as a separated file (Default \code{TRUE}).}

\item{save_bands}{Bands to be saved (Default \code{its.bands}).}
}
\value{
The same input massits predicted tibble.
}
\description{
Save chunks of data to files in GTiff format.
              The prefix of file names is the informed \code{file}.
              This function looks for \code{its_raster} attribute of
              \code{chunk.tb} input.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
