% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.features.R
\name{its.raw.feat}
\alias{its.raw.feat}
\title{massits features functions}
\usage{
its.raw.feat(d, t_length = 23)
}
\arguments{
\item{d}{Any valid set of bands (i.e. named list of vectors)}

\item{t_length}{The length of each data interval to be classified. The length of \code{d} must be
multiple of \code{t_length} (Default \code{23}).}
}
\value{
Massits feature tibble
}
\description{
Transforms a massits tibble as an massits feature tibble
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
