% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.R
\docType{data}
\name{its.sample_cols}
\alias{its.sample_cols}
\title{massits utils globals}
\format{An object of class \code{character} of length 4.}
\usage{
its.sample_cols
}
\description{
Massits tibble column names
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{datasets}
