% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.samples.R
\name{its.scale}
\alias{its.scale}
\title{massits time series functions}
\usage{
its.scale(m, factors = c(1e-04), bands = its.bands())
}
\arguments{
\item{m}{A valid massits tibble}

\item{factors}{Scaling factor to multiply all selected field values
(Default \code{0.0001}).}

\item{bands}{A string vector indicating those fields (bands)
to be re-scaled. All non selected fields
will remain unchanged (Default \code{its.bands()}).}
}
\value{
Massits table
}
\description{
Scaling factor to transform values of massits tibble's fields.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
