% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.utils.R
\name{its.select}
\alias{its.select}
\title{massits utils functions}
\usage{
its.select(m = NULL, ...)
}
\arguments{
\item{m}{A valid massits tibble}

\item{...}{Bands names arguments to return.}
}
\value{
A massits tibble corresponding with selected bands.
}
\description{
Select bands of a massits \code{m} data.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
