% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.utils.R
\name{its.summary}
\alias{its.summary}
\title{massits utils functions}
\usage{
its.summary(m)
}
\arguments{
\item{m}{A valid massits tibble}
}
\value{
Summary tibble
}
\description{
Return a summary of massits data input.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
