% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.time_break.R
\name{its.t_break.count}
\alias{its.t_break.count}
\title{massits break functions}
\usage{
its.t_break.count(m, time_break)
}
\arguments{
\item{m}{A valid massits tibble}

\item{time_break}{A numeric vector indicating sequences of time series to be breaked.
This vector can be computed in \code{its.t_break()} function.}
}
\value{
Table matrix
}
\description{
Counts the number of measures in each distinct pair of
              \code{sample_id} and \code{time_break}.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
