% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.time_break.R
\name{its.t_break}
\alias{its.t_break}
\title{massits break functions}
\usage{
its.t_break(t_start, t_interval, m = NULL)
}
\arguments{
\item{t_start}{A valid time series index. Can be \code{numeric} or \code{lubridate::date}}

\item{t_interval}{A valid interval to compose a theoretical time series breaking periods
Can be numeric if \code{from} is numeric or a \code{lubridate::period}
if \code{from} is \code{datetime}.}

\item{m}{A valid massits tibble}
}
\value{
Numeric vector
}
\description{
Breaks the time series into segments regarding to \code{t_start} and
              \code{t_interval} parameters. A column \code{time_break} is
              created/updated with the identifier of corresponding time_series's segment.
              The massits tibble is then returned.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
