% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.R
\name{its.translate}
\alias{its.translate}
\title{massits time series functions}
\usage{
its.translate(m, amounts = c(3), bands = its.bands())
}
\arguments{
\item{m}{A valid massits tibble}

\item{amounts}{Amount value to be added in all selected field values.
(Default \code{3}).}

\item{bands}{A string vector indicating those fields (bands)
to be displaced. All non selected fields
will remain unchanged (Default \code{its.bands()}).}
}
\value{
Massits table
}
\description{
Translation to be added in values of massits tibble's fields.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
