% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.ml.R
\name{its.tune}
\alias{its.tune}
\title{massits machine learning functions}
\usage{
its.tune(f, ml_model = its.ml.model.svm_radial(), range, cross = 5)
}
\arguments{
\item{f}{A valid massits features tibble}

\item{ml_model}{A valid machine learning method function}

\item{range}{A named list to be expanded that is passed as arguments to method function}

\item{cross}{Cross validation partitions to be test for each parameter combinations}
}
\value{
Tibble of parameters and accuracy
}
\description{
Proceeds a tunning of parameters for a given machine learning method
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
