% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.utils.R
\name{its.valid}
\alias{its.valid}
\title{massits utils functions}
\usage{
its.valid(m, err_desc = NULL)
}
\arguments{
\item{m}{A valid massits tibble}

\item{err_desc}{An error description to be returned.}
}
\value{
Logical
}
\description{
Tests if \code{m} is a valid massits data.
              Returns \code{TRUE} if the input is valid. Else, returns \code{FALSE} or
              throws an exception if \code{err_desc} is informed.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
