% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.ml.R
\name{its.ml.cross_validation}
\alias{its.ml.cross_validation}
\title{massits machine learning functions}
\usage{
its.ml.cross_validation(f, ml_model = its.ml.model.svm_radial(), cross = 5)
}
\arguments{
\item{f}{A valid massits features tibble}

\item{ml_model}{A valid machine learning model function}

\item{cross}{Cross validation partitions to be test for each parameter combinations}
}
\value{
A confusion matrix returned by \code{\link{confusionMatrix}} of \code{caret} package
}
\description{
Proceeds a cross validation for a given machine learning method.
              The number of partitions is given by \code{cross} parameter.
              The result is a tibble with a "reference" and "predicted" columns
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
