% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.R
\name{its}
\alias{its}
\title{massits time series functions}
\usage{
its(x, col_names = NULL)
}
\arguments{
\item{x}{A data source to be converted to massits tibble}

\item{col_names}{A string vector informing corresponding valid column names of the input
as expected in a massive data tibble}
}
\value{
Massits tibble
}
\description{
Create a new massits tibble loaded from \code{x} input.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
