% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its.utils.R
\name{its.summary}
\alias{its.summary}
\alias{its.summary}
\alias{.its.factory}
\title{massits utils functions}
\usage{
its.summary(m)

.its.factory(x = NULL, its_fun)
}
\arguments{
\item{m}{A valid massits tibble}

\item{x}{A valid massits or massits features tibble}

\item{its_fun}{A function that receives as argument an massitis (features) tibble}
}
\value{
Summary tibble

Function evaluation or an enclosed function
}
\description{
Return a summary of massits data input.

Return an evaluation of a given function or an enclosure.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
